% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemut.R
\name{generate_random_tree}
\alias{generate_random_tree}
\title{Generate a random tree}
\usage{
generate_random_tree(nsamples)
}
\arguments{
\item{nsamples}{Number of tips}
}
\value{
phylo object
}
\description{
Generates a random tree with the specified number of tips. For a more realistic tree use e.g. rsimpop
}
