% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemut.R
\name{reconstruct_genotype_summary}
\alias{reconstruct_genotype_summary}
\title{Constructs a genotype summary.}
\usage{
reconstruct_genotype_summary(phylo)
}
\arguments{
\item{phylo}{An APE tree}
}
\value{
a list with two elements: df - a data.frame where each row represents a branch; samples - a character vector of tip labels.
}
\description{
The summary is represented as a data.frame where each row corresponds to a branch.
}
