% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemut.R
\name{simulate_reads_from_tree}
\alias{simulate_reads_from_tree}
\title{Simulates read counts from the specfified genotype summary using a binomial model}
\usage{
simulate_reads_from_tree(
  df,
  avgdepth,
  n_artifacts = 0,
  p.error = 0.01,
  vaf = 0.5
)
}
\arguments{
\item{df}{list. genotype summary}

\item{n_artifacts}{Number of artifact to include}

\item{p.error}{Error rate}

\item{vaf}{VAF for determining mutant read probability.}

\item{avgdeph}{Average depth}

\item{depth}{Depth count matrix. Each row is a mutation and each column is a clonal sample}
}
\value{
Returns a list containing:(mtr=mtr,depth=depth,edge=ml,df=df,p.error=p.error
}
\description{
Simulates read counts from the specfified genotype summary using a binomial model
}
